Use Whiteboard pattern
update the http-tools module to use the whiteboard pattern instead of the deprecated HttpService and ExtHttpService

Update tests and reactivate them

+++ platform/northbound/http-tools/src/test/java/org/eclipse/sensinact/gateway/nthbnd/forward/test/TestForwardingService.java
@@ -10,20 +10,23 @@
  */
 package org.eclipse.sensinact.gateway.nthbnd.forward.test;
 
+import static org.junit.Assert.assertEquals;
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.Map;
+
 import org.eclipse.sensinact.gateway.common.bundle.Mediator;
 import org.eclipse.sensinact.gateway.test.MidOSGiTest;
+import org.eclipse.sensinact.gateway.util.IOUtils;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.Test;
 import org.osgi.framework.Bundle;
 import org.skyscreamer.jsonassert.JSONAssert;
-
-import java.io.File;
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.util.Map;
-
-import static org.junit.Assert.assertEquals;
 
 @SuppressWarnings({"unchecked", "rawtypes"})
 public class TestForwardingService extends MidOSGiTest {
@@ -68,24 +71,72 @@ public class TestForwardingService extends MidOSGiTest {
      */
     @Override
     protected void doInit(Map configuration) {
-        configuration.put("felix.auto.start.1", "file:target/felix/bundle/org.osgi.compendium.jar " + "file:target/felix/bundle/org.apache.felix.configadmin.jar " + "file:target/felix/bundle/org.apache.felix.framework.security.jar ");
-        configuration.put("felix.auto.install.2", "file:target/felix/bundle/sensinact-utils.jar " + "file:target/felix/bundle/sensinact-common.jar " + "file:target/felix/bundle/sensinact-datastore-api.jar " + "file:target/felix/bundle/sensinact-framework-extension.jar " + "file:target/felix/bundle/sensinact-security-none.jar " + "file:target/felix/bundle/sensinact-generic.jar");
 
-        configuration.put("felix.auto.start.2", "file:target/felix/bundle/sensinact-test-configuration.jar " + "file:target/felix/bundle/sensinact-signature-validator.jar " + "file:target/felix/bundle/sensinact-core.jar ");
-        configuration.put("felix.auto.start.3", "file:target/felix/bundle/javax.servlet-api.jar " + "file:target/felix/bundle/org.apache.felix.http.api.jar " + "file:target/felix/bundle/org.apache.felix.http.jetty.jar " + "file:target/felix/bundle/http.jar " + "file:target/felix/bundle/sensinact-northbound-access.jar " + "file:target/felix/bundle/rest-access.jar " + "file:target/felix/bundle/dynamicBundle.jar");
-        configuration.put("felix.auto.start.4", "file:target/felix/bundle/slider.jar " + "file:target/felix/bundle/light.jar ");
-        configuration.put("felix.log.level", "4");
+        configuration.put("felix.auto.start.1",  
+                "file:target/felix/bundle/org.osgi.service.component.jar "+  
+                "file:target/felix/bundle/org.osgi.service.cm.jar "+  
+                "file:target/felix/bundle/org.osgi.service.metatype.jar "+  
+                "file:target/felix/bundle/org.osgi.namespace.extender.jar "+  
+                "file:target/felix/bundle/org.osgi.util.promise.jar "+  
+                "file:target/felix/bundle/org.osgi.util.function.jar "+  
+                "file:target/felix/bundle/org.osgi.service.log.jar "  +
+                "file:target/felix/bundle/org.apache.felix.log.jar " + 
+                "file:target/felix/bundle/org.apache.felix.scr.jar " +
+        		"file:target/felix/bundle/org.apache.felix.fileinstall.jar " +
+        		"file:target/felix/bundle/org.apache.felix.configadmin.jar " + 
+        		"file:target/felix/bundle/org.apache.felix.framework.security.jar ");
+        configuration.put("felix.auto.install.2",  
+        	    "file:target/felix/bundle/org.eclipse.paho.client.mqttv3.jar " + 
+                "file:target/felix/bundle/mqtt-utils.jar " + 
+        	    "file:target/felix/bundle/sensinact-utils.jar " + 
+                "file:target/felix/bundle/sensinact-common.jar " + 
+        	    "file:target/felix/bundle/sensinact-datastore-api.jar " + 
+                "file:target/felix/bundle/sensinact-security-none.jar " + 
+                "file:target/felix/bundle/sensinact-generic.jar " + 
+                "file:target/felix/bundle/slf4j-api.jar " + 
+                "file:target/felix/bundle/slf4j-simple.jar");
+        configuration.put("felix.auto.start.2", 
+        		"file:target/felix/bundle/sensinact-signature-validator.jar " + 
+        		"file:target/felix/bundle/sensinact-core.jar ");
+        configuration.put("felix.auto.start.3", 
+        		"file:target/felix/bundle/org.apache.felix.http.servlet-api.jar " + 
+        		"file:target/felix/bundle/org.apache.felix.http.api.jar "+
+        		"file:target/felix/bundle/org.apache.felix.http.jetty.jar " + 
+        		"file:target/felix/bundle/http.jar " +
+        		"file:target/felix/bundle/sensinact-northbound-access.jar " + 
+                "file:target/felix/bundle/rest-access.jar");
+        configuration.put("felix.auto.start.4", 
+        		"file:target/felix/bundle/dynamicBundle.jar " + 
+                "file:target/felix/bundle/slider.jar " + 
+        		"file:target/felix/bundle/light.jar ");
         configuration.put("org.eclipse.sensinact.gateway.security.jks.filename", "target/felix/bundle/keystore.jks");
         configuration.put("org.eclipse.sensinact.gateway.security.jks.password", "sensiNact_team");
+
+        configuration.put("org.eclipse.sensinact.gateway.location.latitude", "45.2d");
+        configuration.put("org.eclipse.sensinact.gateway.location.longitude", "5.7d");
+
         configuration.put("org.osgi.service.http.port", "8898");
         configuration.put("org.apache.felix.http.jettyEnabled", "true");
+        configuration.put("org.apache.felix.http.whiteboardEnabled", "true");
+
+        try {
+        	String fileName = "sensinact.config";
+            File testFile = new File(new File("src/test/resources"), fileName);
+            URL testFileURL = testFile.toURI().toURL();
+            FileOutputStream output = new FileOutputStream(new File(loadDir,fileName));
+            byte[] testCng = IOUtils.read(testFileURL.openStream(), true);
+            IOUtils.write(testCng, output);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 
     @Test
     public void testForwarding() throws Exception {
+        Thread.sleep(5000);
         Mediator mediator = new Mediator(context);
         this.initializeMoke(new File("./extra-src/test/resources/MANIFEST.MF"), new File("./extra-src/test/resources/meta"), new File("./target/extra-test-classes"));
-
+        
         String simulated1 = HttpServiceTestClient.newRequest(mediator, HTTP_ROOTURL + "/sensinact/providers", null, "GET");
         System.out.println(simulated1);
 
@@ -105,6 +156,7 @@ public class TestForwardingService extends MidOSGiTest {
 
     @Test
     public void testCallback() throws Exception {
+        Thread.sleep(5000);
         Mediator mediator = new Mediator(context);
         this.initializeMoke(new File("./extra-src2/test/resources/MANIFEST.MF"), new File("./extra-src2/test/resources/meta"), new File("./target/extra-test-classes2"));
         try {
@@ -112,8 +164,8 @@ public class TestForwardingService extends MidOSGiTest {
 
             System.out.println(simulated1);
             assertEquals("[GET]/callbackTest1", simulated1);
+            
             String simulated2 = HttpServiceTestClient.newRequest(mediator, HTTP_ROOTURL + "/callbackTest1/withContent", "MyContent", "POST");
-
             System.out.println(simulated2);
             assertEquals("[POST]/callbackTest1/withContent", simulated2);
         } catch (Exception e) {
@@ -123,8 +175,11 @@ public class TestForwardingService extends MidOSGiTest {
 
     private void initializeMoke(File manifestFile, File... sourceDirectories) throws Exception {
         File tmpDirectory = new File("./target/felix/tmp");
-        new File(tmpDirectory, "dynamicBundle.jar").delete();
-
+        if(!tmpDirectory.exists()) {
+        	tmpDirectory.mkdir();
+        } else {
+        	new File(tmpDirectory, "dynamicBundle.jar").delete();
+        }
         int length = (sourceDirectories == null ? 0 : sourceDirectories.length);
         File[] sources = new File[length + 1];
         int index = 0;
@@ -140,13 +195,11 @@ public class TestForwardingService extends MidOSGiTest {
 
         ClassLoader current = Thread.currentThread().getContextClassLoader();
         Thread.currentThread().setContextClassLoader(super.classloader);
+        Thread.sleep(5000);
         try {
             bundle.start();
-
         } catch (Exception e) {
-
             e.printStackTrace();
-
         } finally {
             Thread.currentThread().setContextClassLoader(current);
         }
